rpc interface: remove adapters, directly use controllers
The added complexity of the RPC message injectors doesn't outweigh
the gains (less duplicated code between users of the controller).

This commit removes the adapter traits, if we later decide we need them
we can revert this individual commit.
